separate indexing logic (all tests pass)
